fix CC and CXX rbconfig values for yjit builds
the final change (removing CXX=false) is only necessary for builds
against glibc < 2.17. since yjit requires glibc newer than that,
the inreplace was no longer needed, and started to throw an error.
disabling the inreplace resolved that error, but re-created the problem
that led to the replacement in the first place: trying to find the exact
gcc executable that the ruby had been compiled against (eg `gcc-12`) and
then failing to compile anything with a native extension, because it
didn't exist on the machine where the gem was being installed.