stop static-linking gem extensions
this seems to be the underlying cause of #80, where the functions
provided by the io-console gem were both linked statically into the ruby
binary itself, and also loaded dynamically at runtime when the gem's
compiled .so was loaded. without this flag, it's possible to install and
load the existing io-console gem in eg `rails console` without any
errors.