use macos 15 intel
GitHub has chosen, for some reason, to refuse to support macOS 14 on
Intel. You can have macOS 15 on Intel, which is 14 months old, or you
can have macOS 13 on Intel, which they will delete in two weeks.

I guess this means GitHub's N-1 policy ensures we only support the
current OS and the single OS before that, which seems like not enough
time to me.