only use setup-ruby for Ruby 3.2 that needs it
we originally added setup-ruby because building the various Ruby 3.2
versions was failing if the baseruby was not also Ruby 3.2.

right now, trying to run setup-ruby on the exact same ruby is causing
the 3.4.8 builds to fail, because setup-ruby doesnt have 3.4.8 available
yet. hopefully this will make all the versions work. maybe.